+++ src/protocol/V4_2.php
@@ -10,6 +10,18 @@ namespace Bolt\protocol;
  * @see https://7687.org/bolt/bolt-protocol-message-specification-4.html#version-42
  * @package Bolt\protocol
  */
-class V4_2 extends V4_1
+class V4_2 extends AProtocol
 {
+    use \Bolt\protocol\v1\ResetMessage;
+
+    use \Bolt\protocol\v3\RunMessage;
+    use \Bolt\protocol\v3\BeginMessage;
+    use \Bolt\protocol\v3\CommitMessage;
+    use \Bolt\protocol\v3\RollbackMessage;
+    use \Bolt\protocol\v3\GoodbyeMessage;
+
+    use \Bolt\protocol\v4\PullMessage;
+    use \Bolt\protocol\v4\DiscardMessage;
+
+    use \Bolt\protocol\v4_1\HelloMessage;
 }
